refactor Flow editor into focused modules
