+++ apps/desktop/src/renderer/src/pages/studio/flow-editor/flow-canvas-model.test.ts
@@ -1,116 +1,29 @@
-import { createElement } from "react";
-import { renderToStaticMarkup } from "react-dom/server";
 import { describe, expect, it } from "vitest";
 
-import type { PragmaFlowPrompt, PragmaFlowResource, PragmaResource } from "@pragma/interpreter/ast";
-
-import type { PragmaProjectSnapshot } from "../../../../../shared/desktop-api.ts";
-import { newSchemaField, objectSchemaToFields } from "../JsonSchemaFieldsEditor.tsx";
 import { createEmptyFlow } from "./flow-model.ts";
 import {
-  buildCanvasNodes,
   buildCanvasEdges,
+  buildCanvasNodes,
   canvasPositions,
   createRouteTransition,
   defaultStep,
-  END_NODE_ID,
-  FAIL_NODE_ID,
-  FLOW_ERROR_AUTO_DISMISS_MS,
-  FlowEditor,
-  FlowTimeoutField,
-  flowTimeoutMilliseconds,
-  flowTimeoutUnit,
-  flowTimeoutValue,
-  flowRuntimeProfile,
-  flowVariableOptions,
   inspectorNodeId,
   nextAvailableNodePosition,
-  nextHumanOptionNumber,
   normalizeConnectionDestination,
-  normalizePromptSegments,
-  promptSegmentsFromEditorNodes,
-  PromptTemplateEditor,
   rebuildCanvasNodesPreservingSelection,
   removeEdgeFromFlow,
-  removeHumanOption,
   resourceTargets,
   routeFieldOptions,
-  RuntimeBindingEditor,
-  START_NODE_ID,
-  validateFlowRuntimeSelections,
-} from "./FlowEditor.tsx";
+} from "./flow-canvas-model.ts";
+import { END_NODE_ID, FAIL_NODE_ID, START_NODE_ID } from "./flow-canvas-types.ts";
+import { flowVariableOptions } from "./flow-prompt-editor.tsx";
+import { flowFixture } from "./flow-editor-test-fixtures.ts";
 
-describe("Flow editor canvas", () => {
-  it("dismisses transient editor errors after five seconds", () => {
-    expect(FLOW_ERROR_AUTO_DISMISS_MS).toBe(5_000);
-  });
-
-  it("edits Flow timeouts in human-scale units while preserving milliseconds in the DSL", () => {
-    expect(flowTimeoutUnit(undefined)).toBe("hours");
-    expect(flowTimeoutUnit(90_000)).toBe("seconds");
-    expect(flowTimeoutUnit(120_000)).toBe("minutes");
-    expect(flowTimeoutUnit(7_200_000)).toBe("hours");
-    expect(flowTimeoutUnit(172_800_000)).toBe("days");
-
-    expect(flowTimeoutValue(7_200_000, "hours")).toBe(2);
-    expect(flowTimeoutMilliseconds(30, "seconds")).toBe(30_000);
-    expect(flowTimeoutMilliseconds(15, "minutes")).toBe(900_000);
-    expect(flowTimeoutMilliseconds(2, "hours")).toBe(7_200_000);
-    expect(flowTimeoutMilliseconds(3, "days")).toBe(259_200_000);
-    expect(flowTimeoutMilliseconds(0, "seconds")).toBeUndefined();
-
-    const unlimitedHtml = renderToStaticMarkup(
-      createElement(FlowTimeoutField, {
-        timeoutMs: undefined,
-        onChange: () => undefined,
-      }),
-    );
-    expect(unlimitedHtml).toContain("Never expires");
-    expect(unlimitedHtml).toContain('type="checkbox" checked=""');
-    expect(unlimitedHtml).not.toContain("Timeout (ms)");
-
-    const finiteHtml = renderToStaticMarkup(
-      createElement(FlowTimeoutField, {
-        timeoutMs: 7_200_000,
-        onChange: () => undefined,
-      }),
-    );
-    expect(finiteHtml).toContain('value="2"');
-    expect(finiteHtml).toContain('<option value="hours" selected="">Hours</option>');
-  });
-
-  it("exposes palette items as drag-only controls", () => {
-    const project: PragmaProjectSnapshot = {
-      schemaVersion: "pragma.project-snapshot/v3",
-      projectId: "test-project",
-      revision: 0,
-      resources: [],
-      diagnostics: [],
-    };
-    const html = renderToStaticMarkup(
-      createElement(FlowEditor, {
-        project,
-        error: null,
-        onCancel: () => undefined,
-        onSave: async () => true,
-      }),
-    );
-
-    expect(html).toContain('class="flow-palette-item is-expert" draggable="true"');
-    expect(html).not.toContain('<button class="flow-palette-item');
-    expect(html).toContain("Drag to canvas");
-    expect(html).not.toContain("press Enter");
-    expect(html).not.toContain("flow-palette-item is-action");
-  });
-
+describe("flow-canvas-model", () => {
   it("offers the built-in Expert as a Flow target outside project resources", () => {
     expect(
-      resourceTargets([], "0000000000000001", [
-        { ref: "expert:0000000000pragma", name: "Pragma" },
-      ]),
-    ).toEqual([
-      { kind: "expert", ref: "expert:0000000000pragma", label: "Pragma" },
-    ]);
+      resourceTargets([], "0000000000000001", [{ ref: "expert:0000000000pragma", name: "Pragma" }]),
+    ).toEqual([{ kind: "expert", ref: "expert:0000000000pragma", label: "Pragma" }]);
   });
 
   it("leaves a new semantic step unselected when no matching resource exists", () => {
@@ -195,21 +108,6 @@ describe("Flow editor canvas", () => {
 
     expect(terminalResultNodes.find((node) => node.id === END_NODE_ID)?.selected).toBe(true);
     expect(mappedResultNodes.find((node) => node.id === END_NODE_ID)?.selected).toBe(true);
-  });
-
-  it("allows every Human option to be removed while keeping new values unique", () => {
-    const initial = [
-      { value: "option_1", label: "Option 1" },
-      { value: "option_2", label: "Option 2" },
-    ];
-
-    const oneOption = removeHumanOption(initial, 0);
-    const noOptions = removeHumanOption(oneOption, 0);
-
-    expect(oneOption).toEqual([{ value: "option_2", label: "Option 2" }]);
-    expect(noOptions).toEqual([]);
-    expect(nextHumanOptionNumber(oneOption)).toBe(1);
-    expect(nextHumanOptionNumber(initial)).toBe(3);
   });
 
   it("places repeatedly added nodes into distinct non-overlapping grid cells", () => {
@@ -511,300 +409,4 @@ describe("Flow editor canvas", () => {
 
     expect(after.find((node) => node.id === "review")?.selected).toBe(true);
   });
-
-  it("derives native and structured variables and marks branch-only output optional", () => {
-    const flow = flowFixture();
-    flow.spec.graph.start = "start";
-    flow.spec.graph.steps = {
-      start: {
-        expert: { ref: "expert:caadt9e550f04adk" },
-        prompt: { segments: [{ text: "Start" }] },
-      },
-      scored: {
-        expert: { ref: "expert:p176qzwdwbj85253" },
-        prompt: { segments: [{ text: "Score" }] },
-        output: {
-          schema: {
-            type: "object",
-            properties: { score: { type: "number" }, note: { type: "string" } },
-            required: ["score"],
-            additionalProperties: false,
-          },
-        },
-      },
-      alternate: {
-        expert: { ref: "expert:q4k4hz54yzem8ktn" },
-        prompt: { segments: [{ text: "Alternate" }] },
-      },
-      join: {
-        expert: { ref: "expert:zz8emaxka43mvhd1" },
-        prompt: { segments: [{ text: "Join" }] },
-      },
-    };
-    flow.spec.graph.transitions = {
-      start: {
-        route: "branch",
-        cases: { scored: "scored", alternate: "alternate" },
-      },
-      scored: "join",
-      alternate: "join",
-      join: { end: true },
-    };
-
-    const options = flowVariableOptions(flow, "join");
-    expect(options).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({ label: "start.result", optional: false }),
-        expect.objectContaining({ label: "scored.result", optional: true }),
-        expect.objectContaining({ label: "scored.result.score", optional: true }),
-        expect.objectContaining({ label: "alternate.result", optional: true }),
-      ]),
-    );
-    expect(options.some((option) => option.label === "alternate.result.anything")).toBe(false);
-  });
-
-  it("preserves structured-output field identity when the parent echoes a cloned schema", () => {
-    const field = {
-      ...newSchemaField(),
-      name: "summary",
-      value: {
-        type: "object" as const,
-        fields: [{ ...newSchemaField(), name: "details" }],
-      },
-    };
-    const schema = {
-      type: "object" as const,
-      properties: {
-        summary: {
-          type: "object" as const,
-          properties: { details: { type: "string" as const } },
-          required: ["details"],
-          additionalProperties: false as const,
-        },
-      },
-      required: ["summary"],
-      additionalProperties: false as const,
-    };
-
-    const echoed = objectSchemaToFields(structuredClone(schema), [field]);
-
-    expect(echoed[0]?.id).toBe(field.id);
-    expect(echoed[0]?.value.fields[0]?.id).toBe(field.value.fields[0]?.id);
-  });
-
-  it("renders prompt variables inline inside one editor instead of creating textareas", () => {
-    const flow = createEmptyFlow("prompt_editor");
-    flow.spec.graph.start = "writer";
-    flow.spec.graph.steps.writer = {
-      expert: { ref: "expert:1xddvess309a6gme" },
-      prompt: {
-        segments: [
-          { text: "Use " },
-          { variable: { source: "flow-input", path: [] } },
-          { text: " and " },
-          { variable: { source: "flow-input", path: [] } },
-          { text: "." },
-        ],
-      },
-    };
-    flow.spec.graph.transitions.writer = { end: true };
-
-    const html = renderToStaticMarkup(
-      createElement(PromptTemplateEditor, {
-        flow,
-        stepId: "writer",
-        value: flow.spec.graph.steps.writer!.prompt,
-        onChange: () => undefined,
-      }),
-    );
-
-    expect(html.match(/class="flow-prompt-editor"/g)).toHaveLength(1);
-    expect(html.match(/class="flow-variable-chip"/g)).toHaveLength(2);
-    expect(html).not.toContain("<textarea");
-    expect(html).toContain('contentEditable="true"');
-    expect(html).toContain('<div class="flow-inspector-field" role="group"');
-    expect(html).not.toContain('<label class="flow-inspector-field"><span>Prompt</span>');
-  });
-
-  it("merges adjacent prompt text while preserving inline variable order", () => {
-    const variable = { source: "flow-input" as const, path: ["goal"] };
-    const segments: PragmaFlowPrompt["segments"] = [
-      { text: "Review " },
-      { text: "this: " },
-      { variable },
-      { text: "\nCarefully." },
-    ];
-
-    expect(normalizePromptSegments(segments)).toEqual([
-      { text: "Review this: " },
-      { variable },
-      { text: "\nCarefully." },
-    ]);
-  });
-
-  it("preserves a variable inserted inside an existing prompt text node", () => {
-    const variable = { source: "flow-input" as const, path: ["goal"] };
-    const textNode = (text: string) =>
-      ({
-        nodeType: 3,
-        textContent: text,
-        childNodes: [],
-      }) as unknown as Node;
-    const elementNode = (
-      children: readonly Node[],
-      encodedVariable?: PragmaFlowPrompt["segments"][number],
-    ) =>
-      ({
-        nodeType: 1,
-        textContent: null,
-        childNodes: children,
-        dataset:
-          encodedVariable !== undefined && "variable" in encodedVariable
-            ? { flowVariable: encodeURIComponent(JSON.stringify(encodedVariable.variable)) }
-            : {},
-      }) as unknown as Node;
-
-    const nestedVariableChip = elementNode([textNode("Flow input.goal"), textNode("×")], {
-      variable,
-    });
-    const textSpan = elementNode([
-      textNode("Review "),
-      nestedVariableChip,
-      textNode(" carefully."),
-    ]);
-
-    expect(promptSegmentsFromEditorNodes([textSpan])).toEqual([
-      { text: "Review " },
-      { variable },
-      { text: " carefully." },
-    ]);
-  });
-
-  it("shows actual Runtime environments and requires a model for a Runtime override", () => {
-    const resources = runtimeResources();
-    const runtime = {
-      id: "codex-local",
-      isDefault: true,
-      kind: "codex",
-      displayName: "Codex Local",
-      status: "available" as const,
-      models: [
-        {
-          id: "gpt-5.6",
-          displayName: "GPT-5.6",
-          provider: {
-            kind: "runtime-managed" as const,
-            id: "openai",
-            displayName: "OpenAI",
-          },
-        },
-      ],
-    };
-    const html = renderToStaticMarkup(
-      createElement(RuntimeBindingEditor, {
-        value: undefined,
-        allowModel: true,
-        targetKind: "expert",
-        targetRef: "expert:1xddvess309a6gme",
-        resources,
-        runtimes: [runtime],
-        onSupportingResource: () => undefined,
-        onChange: () => undefined,
-      }),
-    );
-
-    expect(html).toContain("Codex Local");
-    expect(html).not.toContain("Writer Runtime");
-
-    const generated = flowRuntimeProfile(runtime);
-    const flow = createEmptyFlow("7g0mkg5w480wvfgt");
-    flow.spec.graph.start = "writer";
-    flow.spec.graph.steps.writer = {
-      expert: { ref: "expert:1xddvess309a6gme" },
-      runtime: { ref: `runtime-profile:${generated.metadata.id}` },
-    };
-    flow.spec.graph.transitions.writer = { end: true };
-
-    expect(validateFlowRuntimeSelections(flow, [...resources, generated])).toEqual([
-      expect.objectContaining({
-        stepId: "writer",
-        message: "Choose a model when overriding the node Runtime.",
-      }),
-    ]);
-
-    flow.spec.graph.steps.writer!.runtime!.modelSelection = {
-      model: { providerId: "openai", modelId: "gpt-5.6" },
-    };
-    expect(validateFlowRuntimeSelections(flow, [...resources, generated])).toEqual([]);
-  });
 });
-
-function runtimeResources(): PragmaResource[] {
-  return [
-    {
-      apiVersion: "pragma/v3",
-      kind: "RuntimeProfile",
-      metadata: {
-        id: "rdzgnq05qfqcpqcm",
-        name: "Writer Runtime",
-        description: "Writer Runtime",
-        tags: ["desktop-managed"],
-      },
-      spec: {
-        adapter: "pragma.runtime.profile@v1",
-        config: {
-          runtimeId: "codex-local",
-          providerId: "openai",
-          model: "gpt-5.6",
-        },
-      },
-    },
-    {
-      apiVersion: "pragma/v3",
-      kind: "Expert",
-      metadata: {
-        id: "1xddvess309a6gme",
-        name: "Writer",
-        description: "Writes.",
-        tags: [],
-      },
-      spec: {
-        scope: "Write.",
-        instructions: "Write.",
-        runtime: { ref: "runtime-profile:rdzgnq05qfqcpqcm" },
-        capabilities: [],
-        toolApprovals: {},
-        contextStores: [],
-        plugins: [],
-        tools: [],
-      },
-    },
-  ];
-}
-
-function flowFixture(): PragmaFlowResource {
-  return {
-    apiVersion: "pragma/v3",
-    kind: "Flow",
-    metadata: {
-      id: "qj3t30sa520dvfvj",
-      name: "Review flow",
-      description: "Review a change",
-      tags: [],
-    },
-    spec: {
-      limits: { maxNodeVisits: 10 },
-      graph: {
-        start: "review",
-        steps: {
-          review: {
-            expert: { ref: "expert:t9ne4d8njvvxv2ea" },
-            prompt: { segments: [{ text: "Review this change." }] },
-          },
-        },
-        loops: {},
-        transitions: { review: { end: true } },
-      },
-    },
-  };
-}
